Correct Faulty 'to' parsing
The sendgrid library that this connector is using will only accept one
'to' recipient at a time, and only with a very specific format.
Previously, any 'to' string was being parsed into an array, and then
passed to the sendgrid library where it would fail.  That's annoying.

This fixes the parsing logic to ensure that we're only sending one email
at a time, and only sends the first email.  This may be slightly
sub-optimal, but at least it works.  Futher refactoring should allow us
to send a separate email to each recipient.